business_card {pagedown}	R Documentation

Create business cards

Description

This output format is based on an example in the Github repo https://github.com/RelaxedJS/ReLaXed-examples. See https://pagedown.rbind.io/business-card/ for an example.

Usage

business_card(template = pkg_resource("html", "card.html"))

Arguments

template
The path to the Pandoc template to convert Markdown to HTML.

Value

An R Markdown output format.

Examples

pagedown::business_card()

[Package pagedown version 0.23 Index]