find_chrome {pagedown}	R Documentation

Find Google Chrome, Microsoft Edge or Chromium in the system

Description

On Windows, this function tries to find Chrome or Edge from the registry. On macOS, it returns a hard-coded path of Chrome under ‘/Applications’. On Linux, it searches for chromium-browser and google-chrome from the system's PATH variable.

Usage

find_chrome()

Value

A character string.

[Package pagedown version 0.23 Index]